+++ src/main/java/net/detectivekaktus/item/DotcAbilityItem.java
@@ -1,8 +1,11 @@
 package net.detectivekaktus.item;
 
+import net.detectivekaktus.component.DotcComponents;
+import net.detectivekaktus.component.records.ChargeableComponent;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.tags.TagKey;
+import net.minecraft.world.InteractionResult;
 import net.minecraft.world.InteractionResultHolder;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
@@ -27,7 +30,7 @@ public abstract class DotcAbilityItem extends DotcItem implements HasManaCost, H
         var mana = PlayerMana.get(player);
         var hasInfiniteMaterials = player.hasInfiniteMaterials();
         var notEnoughMana = !hasInfiniteMaterials && getManaCost() > mana.getCurrentMana();
-    var invulnerable = target != null && getInvulnerableTag() != null && target.getType().is(getInvulnerableTag());
+        var invulnerable = target != null && getInvulnerableTag() != null && target.getType().is(getInvulnerableTag());
 
         if (level.isClientSide) {
             if (notEnoughMana)
@@ -54,6 +57,17 @@ public abstract class DotcAbilityItem extends DotcItem implements HasManaCost, H
 
         if (notEnoughMana || invulnerable)
             return InteractionResultHolder.fail(stack);
+
+        if (stack.has(DotcComponents.CHARGEABLE_COMPONENT)) {
+            var component = stack.get(DotcComponents.CHARGEABLE_COMPONENT);
+            if (component.charges() == 0)
+                return InteractionResultHolder.fail(stack);
+
+            stack.set(
+                    DotcComponents.CHARGEABLE_COMPONENT,
+                    ChargeableComponent.consumeCharge(component)
+            );
+        }
 
         if (target instanceof Player interactedPlayer)
             CombatManager.addStickCharge(interactedPlayer);