+++ src/main/java/net/detectivekaktus/item/DotcItems.java
@@ -92,6 +92,7 @@ public class DotcItems {
                 DotcTools.ORCHID_MALEVOLENCE,
                 DotcTools.BLOODTHORN,
                 DotcTools.ROD_OF_ATOS,
+                DotcTools.URN_OF_SHADOWS,
 
                 DotcTools.HEART_OF_TARRASQUE,
                 DotcTools.AEON_DISK,