+++ src/main/java/net/detectivekaktus/sound/item/DotcItemSounds.java
@@ -29,6 +29,7 @@ public class DotcItemSounds {
     public static final SoundEvent BLOODTHORN = DotcSounds.register("item_bloodthorn");
     public static final SoundEvent BLOODTHORN_TARGET = DotcSounds.register("item_bloodthorn_target");
     public static final SoundEvent ROD_OF_ATOS = DotcSounds.register("item_rod_of_atos");
+    public static final SoundEvent URN_OF_SHADOWS = DotcSounds.register("item_urn_of_shadows");
 
     public static void initialize() {}
 }